minor refactoring on debug viewer use
